requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
SQLAlchemy==0.8.2
py-bcrypt==0.3
